fix: on-demand api client initialization
- Extract client initialization logic from `utils.py` into a new `client.py` module.
- Refactor API client creation to instantiate clients on-demand safely using a threading lock, resolving unit test failures and potential thread safety issues.
- Cache credentials lazily to avoid redundant metadata server calls.
- Update imports across all tool files to use the new `client` module.
