+++ analytics_mcp/tools/reporting/metadata.py
@@ -19,10 +19,10 @@ from typing import Any, Dict, List
 
 from analytics_mcp.tools.utils import (
     construct_property_rn,
-    create_data_api_client,
     proto_to_dict,
     proto_to_json,
 )
+from analytics_mcp.tools.client import create_data_api_client
 from google.analytics import data_v1alpha, data_v1beta
