+++ analytics_mcp/tools/utils.py
@@ -16,80 +16,7 @@
 
 from typing import Any, Dict
 
-from google.analytics import (
-    admin_v1beta,
-    data_v1beta,
-    admin_v1alpha,
-    data_v1alpha,
-)
-from google.api_core.gapic_v1.client_info import ClientInfo
-from importlib import metadata
-import google.auth
 import proto
-import sys
-import threading
-
-
-def _get_package_version_with_fallback():
-    """Returns the version of the package.
-
-    Falls back to 'unknown' if the version can't be resolved.
-    """
-    try:
-        return metadata.version("analytics-mcp")
-    except:
-        return "unknown"
-
-
-# Client information that adds a custom user agent to all API requests.
-_CLIENT_INFO = ClientInfo(
-    user_agent=f"analytics-mcp/{_get_package_version_with_fallback()}"
-)
-
-# Read-only scope for Analytics Admin API and Analytics Data API.
-_READ_ONLY_ANALYTICS_SCOPE = (
-    "https://www.googleapis.com/auth/analytics.readonly"
-)
-
-# Initialize credentials and clients at the module level (Eager Loading).
-# This ensures they are created on the main thread during import, avoiding
-# deadlocks in worker threads.
-_CREDENTIALS, _ = google.auth.default(scopes=[_READ_ONLY_ANALYTICS_SCOPE])
-
-_ADMIN_CLIENT = admin_v1beta.AnalyticsAdminServiceClient(
-    client_info=_CLIENT_INFO, credentials=_CREDENTIALS
-)
-_DATA_CLIENT = data_v1beta.BetaAnalyticsDataClient(
-    client_info=_CLIENT_INFO, credentials=_CREDENTIALS
-)
-_ADMIN_ALPHA_CLIENT = admin_v1alpha.AnalyticsAdminServiceClient(
-    client_info=_CLIENT_INFO, credentials=_CREDENTIALS
-)
-_DATA_ALPHA_CLIENT = data_v1alpha.AlphaAnalyticsDataClient(
-    client_info=_CLIENT_INFO, credentials=_CREDENTIALS
-)
-
-
-def create_admin_api_client() -> admin_v1beta.AnalyticsAdminServiceClient:
-    """Returns the Google Analytics Admin API client."""
-    return _ADMIN_CLIENT
-
-
-def create_data_api_client() -> data_v1beta.BetaAnalyticsDataClient:
-    """Returns the Google Analytics Data API client."""
-    return _DATA_CLIENT
-
-
-def create_admin_alpha_api_client() -> (
-    admin_v1alpha.AnalyticsAdminServiceClient
-):
-    """Returns the Google Analytics Admin API (alpha) client."""
-    return _ADMIN_ALPHA_CLIENT
-
-
-def create_data_api_alpha_client() -> data_v1alpha.AlphaAnalyticsDataClient:
-    """Returns the Google Analytics Data API (Alpha) client."""
-    return _DATA_ALPHA_CLIENT
 
 
 def construct_property_rn(property_value: int | str) -> str:
